+++ build.sh
@@ -36,6 +36,14 @@ env GOOS=linux GOARCH=amd64 go build -o ../../dist/linux/unlock
 rm rsrc_windows_amd64.syso
 popd
 
+pushd ./commands/dumpsmc
+echo "Building dumpsmc"
+go-winres make --arch amd64 --product-version $1 --file-version $1
+env GOOS=windows GOARCH=amd64 go build -o ../../dist/windows/dumpsmc.exe
+env GOOS=linux GOARCH=amd64 go build -o ../../dist/linux/dumpsmc
+rm rsrc_windows_amd64.syso
+popd
+
 pushd ./commands/hostcaps
 echo "Building hostcaps"
 go-winres make --arch amd64 --product-version $1 --file-version $1